Fixed migrations + livewire product price part 1 + livewire product variations selector part 1 + env changes + routes fix etc.

+++ database/migrations/tenant/2021_09_28_000909_refactor_products_table.php
@@ -14,15 +14,33 @@ class RefactorProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('attributes');
-            $table->dropColumn('choice_options');
-            $table->dropColumn('colors');
-            $table->dropColumn('variations');
-            $table->dropColumn('current_stock'); // DONE: Create current_stock property in Product model
-            $table->dropColumn('low_stock_quantity'); // DONE: Create low_stock_qty property in Product model and move it to ProductStocks table
-            $table->dropColumn('category_id');
-            $table->dropColumn('subcategory_id');
-            $table->dropColumn('subsubcategory_id');
+            if (Schema::hasColumn('products', 'attributes')) {
+                $table->dropColumn('attributes');
+            }
+            if (Schema::hasColumn('products', 'choice_options')) {
+                $table->dropColumn('choice_options');
+            }
+            if (Schema::hasColumn('products', 'colors')) {
+                $table->dropColumn('colors');
+            }
+            if (Schema::hasColumn('products', 'variations')) {
+                $table->dropColumn('variations');
+            }
+            if (Schema::hasColumn('products', 'current_stock')) {
+                $table->dropColumn('current_stock');
+            }
+            if (Schema::hasColumn('products', 'low_stock_quantity')) {
+                $table->dropColumn('low_stock_quantity');
+            }
+            if (Schema::hasColumn('products', 'category_id')) {
+                $table->dropColumn('category_id');
+            }
+            if (Schema::hasColumn('products', 'subcategory_id')) {
+                $table->dropColumn('subcategory_id');
+            }
+            if (Schema::hasColumn('products', 'subsubcategory_id')) {
+                $table->dropColumn('subsubcategory_id');
+            }
 
             $table->float('purchase_price')->nullable()->change();
         });
