+++ database/migrations/tenant/2021_10_15_171326_remove_unnecessary_fields_from_products_table.php
@@ -14,8 +14,12 @@ class RemoveUnnecessaryFieldsFromProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('category_id');
-            $table->dropColumn('variant_product');
+            if (Schema::hasColumn('products', 'category_id')) {
+                $table->dropColumn('category_id');
+            }
+            if (Schema::hasColumn('products', 'variant_product')) {
+                $table->dropColumn('variant_product');
+            }
         });
     }
 
@@ -27,8 +31,7 @@ class RemoveUnnecessaryFieldsFromProductsTable extends Migration
     public function down()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('category_id')->default(1);
-            $table->integer('variant_product')->default(0);
+
         });
     }
 }
